Small Bug Fixes

+++ php/deleteitem.php
@@ -3,10 +3,16 @@ require_once('connection.php');
 session_start();
 if(isset($_POST['deleteitembtn']))
 {
-    
-   $query1="DELETE FROM inventorydb WHERE itemname='".$_POST['itemname']."'and  email ='".$_SESSION['Email']."'";
-   $result1 =  mysqli_query($con,$query1);
-   header("location:../html/homepage.php?Inserted= Data Deleted");
+   if(empty($_POST['itemname']))
+   {
+      header("location:../html/homepage.php?Empty= Please Fill all the fields.");
+   }
+   else
+   {
+      $query1="DELETE FROM inventorydb WHERE itemname='".$_POST['itemname']."'and  email ='".$_SESSION['Email']."'";
+      $result1 =  mysqli_query($con,$query1);
+      header("location:../html/homepage.php?Inserted= Data Deleted");
+   }
 }
